Se completa el registrar empleado, por alguna razon el registrar Fincas sigue fallando

+++ ProductoraHuevos/src/main/java/com/productora/huevos/repository/FincasRepository.java
@@ -3,6 +3,6 @@ package com.productora.huevos.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.productora.huevos.domain.Fincas;
 
-public interface FincasRepository extends JpaRepository<Fincas, Long>{
+public interface FincasRepository extends JpaRepository<Fincas, Integer>{
 
 }
