+++ ProductoraHuevos/src/main/java/com/productora/huevos/repository/ProduccionRepository.java
@@ -3,6 +3,6 @@ package com.productora.huevos.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.productora.huevos.domain.Produccion;
 
-public interface ProduccionRepository  extends JpaRepository<Produccion, Long> {
+public interface ProduccionRepository  extends JpaRepository<Produccion, Integer> {
 
 }
